Add temporary variables to class members
Instead of creating temporary variables on the fly, they are now part of the private members of the Curve class. I assume that memory space will be allocated for the entire life time of the program for these temporary variables. The original thought was to avoid memory conflict when running in parallel. (This was before the subsequent discovery of an R runif call which is probably the main reason why the parallel computation failed.)