Make algorithm for BS97 to return the EM auxiliary function
The previous version returns the marginal likelihood and ran into the issue of numerical under/overflow a lot. For the new version, we return the auxiliary function instead. This implementation serve as a patch so that fitted models using multiple random starting values can be compared based on the auxiliary function. Stopping for convergence is also implemented based on percentage increase in auxiliary function, i.e. relative tolerance.